# src/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"CartItem": {
				"properties": {
					"baseMeal": {
						"type": "string"
					},
					"proteins": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"combo": {
						"type": "string"
					},
					"subtotal": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"subtotal"
				],
				"type": "object",
				"additionalProperties": false
			},
			"OrderDTO": {
				"properties": {
					"id": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"items": {
						"$ref": "#/components/schemas/CartItem"
					},
					"subtotal": {
						"type": "number",
						"format": "double"
					},
					"deliveryFee": {
						"type": "number",
						"format": "double"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "string",
						"enum": [
							"pending",
							"confirmed",
							"delivered"
						]
					},
					"deliveryType": {
						"type": "string",
						"enum": [
							"pickup",
							"delivery"
						]
					},
					"deliveryAddress": {
						"type": "string"
					},
					"pickupLocation": {
						"type": "string"
					},
					"deliveryWindow": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"phoneNumber",
					"items",
					"subtotal",
					"deliveryFee",
					"total",
					"status",
					"deliveryType",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DeliveryArea": {
				"type": "string",
				"enum": [
					"gk",
					"outside-gk"
				],
				"nullable": false
			},
			"NativeDate": {
				"type": "string"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "practice-web",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/main/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Main"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			}
		},
		"/main/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"token": {},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										},
										{
											"properties": {
												"message": {},
												"token": {}
											},
											"required": [
												"token"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Main"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			}
		},
		"/main/cart/add": {
			"post": {
				"operationId": "AddCart",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Main"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			}
		},
		"/main/order/place": {
			"post": {
				"operationId": "Place",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OrderDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Main"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"deliveryAddress": {
										"type": "string"
									},
									"deliveryArea": {
										"$ref": "#/components/schemas/DeliveryArea"
									},
									"deliveryType": {
										"type": "string",
										"enum": [
											"pickup",
											"delivery"
										]
									}
								},
								"required": [
									"deliveryType"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/menu/home": {
			"get": {
				"operationId": "GetHome",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"combos": {
											"items": {
												"properties": {
													"price": {
														"type": "number",
														"format": "double"
													},
													"name": {
														"type": "string"
													}
												},
												"required": [
													"price",
													"name"
												],
												"type": "object"
											},
											"type": "array"
										},
										"proteins": {
											"items": {
												"properties": {
													"price": {
														"type": "number",
														"format": "double"
													},
													"name": {
														"type": "string"
													}
												},
												"required": [
													"price",
													"name"
												],
												"type": "object"
											},
											"type": "array"
										},
										"deliveryInfo": {
											"properties": {
												"note": {
													"type": "string"
												},
												"window": {
													"type": "string"
												}
											},
											"required": [
												"note",
												"window"
											],
											"type": "object"
										},
										"baseMeal": {
											"properties": {
												"currency": {
													"type": "string"
												},
												"price": {
													"type": "number",
													"format": "double"
												},
												"name": {
													"type": "string"
												}
											},
											"required": [
												"currency",
												"price",
												"name"
											],
											"type": "object"
										},
										"orderButtonText": {
											"type": "string"
										},
										"subtext": {
											"type": "string"
										},
										"headline": {
											"type": "string"
										}
									},
									"required": [
										"combos",
										"proteins",
										"deliveryInfo",
										"baseMeal",
										"orderButtonText",
										"subtext",
										"headline"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Menu"
				],
				"security": [],
				"parameters": []
			}
		},
		"/admin/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"token": {},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										},
										{
											"properties": {
												"message": {},
												"token": {}
											},
											"required": [
												"token"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			}
		},
		"/admin/orders": {
			"get": {
				"operationId": "Get",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"properties": {
											"createdAt": {
												"$ref": "#/components/schemas/NativeDate"
											},
											"deliveryWindow": {
												"type": "string"
											},
											"pickupLocation": {
												"type": "string"
											},
											"deliveryAddress": {
												"type": "string"
											},
											"deliveryType": {
												"type": "string",
												"enum": [
													"pickup",
													"delivery"
												]
											},
											"status": {
												"type": "string",
												"enum": [
													"pending",
													"confirmed",
													"delivered"
												]
											},
											"total": {
												"type": "number",
												"format": "double"
											},
											"deliveryFee": {
												"type": "number",
												"format": "double"
											},
											"subtotal": {
												"type": "number",
												"format": "double"
											},
											"items": {},
											"phoneNumber": {
												"type": "string"
											},
											"id": {
												"type": "string"
											}
										},
										"required": [
											"createdAt",
											"deliveryWindow",
											"pickupLocation",
											"deliveryAddress",
											"deliveryType",
											"status",
											"total",
											"deliveryFee",
											"subtotal",
											"items",
											"phoneNumber",
											"id"
										],
										"type": "object"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [],
				"parameters": []
			}
		},
		"/admin/orders/{id}": {
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"createdAt": {
											"$ref": "#/components/schemas/NativeDate"
										},
										"deliveryWindow": {
											"type": "string"
										},
										"pickupLocation": {
											"type": "string"
										},
										"deliveryAddress": {
											"type": "string"
										},
										"deliveryType": {
											"type": "string",
											"enum": [
												"pickup",
												"delivery"
											]
										},
										"status": {
											"type": "string",
											"enum": [
												"pending",
												"confirmed",
												"delivered"
											]
										},
										"total": {
											"type": "number",
											"format": "double"
										},
										"deliveryFee": {
											"type": "number",
											"format": "double"
										},
										"subtotal": {
											"type": "number",
											"format": "double"
										},
										"items": {},
										"phoneNumber": {
											"type": "string"
										},
										"id": {
											"type": "string"
										}
									},
									"required": [
										"createdAt",
										"deliveryWindow",
										"pickupLocation",
										"deliveryAddress",
										"deliveryType",
										"status",
										"total",
										"deliveryFee",
										"subtotal",
										"items",
										"phoneNumber",
										"id"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}